refactor: refactor five change dialog to be generic
